add active effects to all configurations
