+++ module/item/sheets/WitcherArmorSheet.js
@@ -1,11 +1,8 @@
-
-import WitcheActiveEffectConfigurationSheet from "./configurations/WitcherActiveEffectConfigurationSheet.js";
-import WitcherItemSheet from "./WitcherItemSheet.js";
+import WitcherConfigurationSheet from './configurations/WitcherConfigurationSheet.js';
+import WitcherItemSheet from './WitcherItemSheet.js';
 
 export default class WitcherArmorSheet extends WitcherItemSheet {
-
-    configuration = new WitcheActiveEffectConfigurationSheet(this.item);
-
+    configuration = new WitcherConfigurationSheet(this.item);
 
     get template() {
         return `systems/TheWitcherTRPG/templates/sheets/armor-sheet.hbs`;
@@ -15,7 +12,7 @@ export default class WitcherArmorSheet extends WitcherItemSheet {
     getData() {
         const data = super.getData();
 
-        data.config.Availability.WITCHER = "WITCHER.Item.AvailabilityWitcher";
+        data.config.Availability.WITCHER = 'WITCHER.Item.AvailabilityWitcher';
         data.config.type = this.getTypes();
         data.config.armorLocations = this.getArmorLocations();
 
@@ -24,20 +21,20 @@ export default class WitcherArmorSheet extends WitcherItemSheet {
 
     getTypes() {
         return {
-            Light: "WITCHER.Armor.Light",
-            Medium: "WITCHER.Armor.Medium",
-            Heavy: "WITCHER.Armor.Heavy",
-            Natural: "WITCHER.Armor.Natural",
-        }
+            Light: 'WITCHER.Armor.Light',
+            Medium: 'WITCHER.Armor.Medium',
+            Heavy: 'WITCHER.Armor.Heavy',
+            Natural: 'WITCHER.Armor.Natural'
+        };
     }
 
     getArmorLocations() {
         return {
-            Head: "WITCHER.Armor.LocationHead",
-            Torso: "WITCHER.Armor.LocationTorso",
-            Leg: "WITCHER.Armor.LocationLeg",
-            FullCover: "WITCHER.Armor.LocationFull",
-            Shield: "WITCHER.Armor.LocationShield",
-        }
+            Head: 'WITCHER.Armor.LocationHead',
+            Torso: 'WITCHER.Armor.LocationTorso',
+            Leg: 'WITCHER.Armor.LocationLeg',
+            FullCover: 'WITCHER.Armor.LocationFull',
+            Shield: 'WITCHER.Armor.LocationShield'
+        };
     }
 }
