+++ templates/sheets/item/configuration/alchemicalConfiguration.hbs
@@ -1,4 +1,21 @@
 <form class="{{cssClass}}" autocomplete="off">
     <h1>{{localize "WITCHER.Item.Settings.Alchemical"}}</h1>
-    {{>"systems/TheWitcherTRPG/templates/sheets/item/configuration/partials/consumablePropertiesConfiguration.hbs"}}
+    {{! Sheet Tab Navigation }}
+    <nav class="sheet-tabs tabs flexrow" data-group="primary">
+        <a data-tab="general">{{localize "WITCHER.Item.Settings.General"}}</a>
+        <a data-tab="consumable">{{localize "WITCHER.Item.Settings.ConsumableProperties"}}</a>
+        <a data-tab="effects">{{localize "WITCHER.activeEffect.tab"}}</a>
+    </nav>
+
+    {{! Sheet Body }}
+    <section class="sheet-body">
+        <div class="tab" data-group="primary" data-tab="general">
+        </div>
+        <div class="tab" data-group="primary" data-tab="consumable">
+            {{>"systems/TheWitcherTRPG/templates/sheets/item/configuration/partials/consumablePropertiesConfiguration.hbs"}}
+        </div>
+        <div class="tab effects" data-group="primary" data-tab="effects">
+            {{> systems/TheWitcherTRPG/templates/partials/effect-part.hbs}}
+        </div>
+    </section>
 </form>
