+++ templates/sheets/item/configuration/weaponConfiguration.hbs
@@ -1,5 +1,26 @@
 <form class="{{cssClass}}" autocomplete="off">
     <h1>{{localize "WITCHER.Item.Settings.Weapon"}}</h1>
-    {{>"systems/TheWitcherTRPG/templates/sheets/item/configuration/partials/damagePropertiesConfiguration.hbs"}}
-    {{>"systems/TheWitcherTRPG/templates/sheets/item/configuration/partials/defensePropertiesConfiguration.hbs"}}
+    {{! Sheet Tab Navigation }}
+    <nav class="sheet-tabs tabs flexrow" data-group="primary">
+        <a data-tab="general">{{localize "WITCHER.Item.Settings.General"}}</a>
+        <a data-tab="damageProperties">{{localize "WITCHER.Item.Settings.DamageProperties"}}</a>
+        <a data-tab="defenseProperties">{{localize "WITCHER.Item.Settings.DefenseProperties"}}</a>
+        <a data-tab="effects">{{localize "WITCHER.activeEffect.tab"}}</a>
+    </nav>
+
+    {{! Sheet Body }}
+    <section class="sheet-body">
+        <div class="tab" data-group="primary" data-tab="general">
+        </div>
+        <div class="tab" data-group="primary" data-tab="damageProperties">
+            {{>"systems/TheWitcherTRPG/templates/sheets/item/configuration/partials/damagePropertiesConfiguration.hbs"}}
+        </div>
+        <div class="tab" data-group="primary" data-tab="defenseProperties">
+            {{>"systems/TheWitcherTRPG/templates/sheets/item/configuration/partials/defensePropertiesConfiguration.hbs"}}
+        </div>
+        <div class="tab effects" data-group="primary" data-tab="effects">
+            {{> systems/TheWitcherTRPG/templates/partials/effect-part.hbs}}
+        </div>
+    </section>
+
 </form>
